common module for energy update
