+++ common/common_errors/src/lib.rs
@@ -17,4 +17,3 @@ pub static ERROR_NO_TOKEN_TO_MERGE: &[u8] = b"No token to merge";
 pub static ERROR_PAYMENT_FAILED: &[u8] = b"Payment failed";
 pub static ERROR_PERMISSION_DENIED: &[u8] = b"Permission denied";
 pub static ERROR_PARAMETERS: &[u8] = b"Bad parameters";
-pub static ERROR_ENERGY_UPDATE_SAME_WEEK: &[u8] = b"Can update only after claim rewards";
