+++ dex/farm/src/claim_boost_only.rs
@@ -15,6 +15,7 @@ pub trait ClaimBoostOnlyModule:
     + weekly_rewards_splitting::events::WeeklyRewardsSplittingEventsModule
     + weekly_rewards_splitting::global_info::WeeklyRewardsGlobalInfo
     + weekly_rewards_splitting::locked_token_buckets::WeeklyRewardsLockedTokenBucketsModule
+    + weekly_rewards_splitting::update_claim_progress_energy::UpdateClaimProgressEnergyModule
     + energy_query::EnergyQueryModule
     + token_send::TokenSendModule
     + events::EventsModule
