+++ energy-integration/common-modules/weekly-rewards-splitting/src/update_claim_progress_energy.rs
@@ -1,0 +1,82 @@
+elrond_wasm::imports!();
+
+use common_types::Week;
+use energy_query::Energy;
+
+use crate::ClaimProgress;
+
+pub static ERROR_ENERGY_UPDATE_SAME_WEEK: &[u8] = b"Can update only after claim rewards";
+
+#[elrond_wasm::module]
+pub trait UpdateClaimProgressEnergyModule:
+    energy_query::EnergyQueryModule
+    + week_timekeeping::WeekTimekeepingModule
+    + crate::events::WeeklyRewardsSplittingEventsModule
+    + crate::global_info::WeeklyRewardsGlobalInfo
+    + crate::locked_token_buckets::WeeklyRewardsLockedTokenBucketsModule
+{
+    #[endpoint(updateEnergyForUser)]
+    fn update_energy_for_user(&self, user: ManagedAddress) {
+        let current_week = self.get_current_week();
+        let claim_progress = self.current_claim_progress(&user).get();
+        require!(
+            claim_progress.week == current_week,
+            ERROR_ENERGY_UPDATE_SAME_WEEK
+        );
+        self.update_energy_and_progress(&user);
+    }
+
+    fn update_energy_and_progress(&self, caller: &ManagedAddress) {
+        let current_week = self.get_current_week();
+        let current_user_energy = self.get_energy_entry(caller);
+
+        let progress_mapper = self.current_claim_progress(caller);
+        let opt_progress_for_update = if !progress_mapper.is_empty() {
+            Some(progress_mapper.get())
+        } else {
+            None
+        };
+        self.update_user_energy_for_current_week(
+            caller,
+            current_week,
+            &current_user_energy,
+            opt_progress_for_update,
+        );
+
+        progress_mapper.set(&ClaimProgress {
+            week: current_week,
+            energy: current_user_energy,
+        });
+    }
+
+    fn update_user_energy_for_current_week(
+        &self,
+        user: &ManagedAddress,
+        current_week: Week,
+        current_energy: &Energy<Self::Api>,
+        opt_existing_claim_progres: Option<ClaimProgress<Self::Api>>,
+    ) {
+        let (last_active_week, prev_energy) = match opt_existing_claim_progres {
+            Some(existing_claim_progress) => {
+                (existing_claim_progress.week, existing_claim_progress.energy)
+            }
+            None => (0, Energy::default()),
+        };
+
+        self.update_global_amounts_for_current_week(
+            current_week,
+            last_active_week,
+            &prev_energy,
+            current_energy,
+        );
+
+        self.emit_update_user_energy_event(user, current_week, current_energy);
+    }
+
+    #[view(getCurrentClaimProgress)]
+    #[storage_mapper("currentClaimProgress")]
+    fn current_claim_progress(
+        &self,
+        user: &ManagedAddress,
+    ) -> SingleValueMapper<ClaimProgress<Self::Api>>;
+}
