+++ energy-integration/farm-boosted-yields/src/lib.rs
@@ -47,6 +47,7 @@ pub trait FarmBoostedYieldsModule:
     + weekly_rewards_splitting::events::WeeklyRewardsSplittingEventsModule
     + weekly_rewards_splitting::global_info::WeeklyRewardsGlobalInfo
     + weekly_rewards_splitting::locked_token_buckets::WeeklyRewardsLockedTokenBucketsModule
+    + weekly_rewards_splitting::update_claim_progress_energy::UpdateClaimProgressEnergyModule
     + energy_query::EnergyQueryModule
 {
     #[endpoint(setBoostedYieldsRewardsPercentage)]
